[FIX] 전체 컴포넌트 구조 변경중
