+++ src/components/Card.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 import background from "../img/loba.svg"
 import kakaobtn from "../img/kakao_login_large_wide.png"
