+++ src/components/dashboard/sidebar/SidebarIcon/Settings.js
